+++ src/main/java/com/bol/kalaha/repository/GameRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.bol.kalaha.model.Game;
 
-public interface GameRepository extends CrudRepository<Game, Integer> {
+public interface GameRepository extends CrudRepository<Game, Long> {
 
-	Optional<Game> findById(long id);
+	Optional<Game> findById(Long id);
 	
 	
 }
